Move reused button row styles to shared-styles
