+++ context/app/static/js/components/Detail/SampleSpecificDatasetsTable/SampleSpecificDatasetsTable.jsx
@@ -12,7 +12,7 @@ import { StyledTableContainer, HeaderCell } from 'js/shared-styles/Table';
 import SectionHeader from 'js/shared-styles/sections/SectionHeader';
 import SectionContainer from 'js/shared-styles/sections/SectionContainer';
 import { LightBlueLink } from 'js/shared-styles/Links';
-import { StyledButtonRow, BottomAlignedTypography } from './style';
+import { StyledButtonRow, BottomAlignedTypography } from 'js/shared-styles/sections/RightAlignedButtonRow';
 
 function SampleSpecificDatasetsTable({ datasets, uuid }) {
   const columns = [
