+++ context/app/static/js/components/savedLists/SavedEntitiesTable/style.js
@@ -1,9 +1,7 @@
 import styled, { css } from 'styled-components';
 import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
 
 import { WhiteBackgroundIconButton } from 'js/shared-styles/buttons';
-import RightAlignedButtonRow from 'js/shared-styles/sections/RightAlignedButtonRow';
 
 const buttonSpacing = css`
   margin-left: ${(props) => props.theme.spacing(1)}px;
@@ -23,13 +21,4 @@ const LeftMarginIconButton = styled(WhiteBackgroundIconButton)`
   ${buttonSpacing}
 `;
 
-const StyledButtonRow = styled(RightAlignedButtonRow)`
-  margin-bottom: ${(props) => props.theme.spacing(1)}px;
-  min-height: 40px;
-`;
-
-const BottomAlignedTypography = styled(Typography)`
-  align-self: flex-end;
-`;
-
-export { Flex, LeftMarginButton, LeftMarginIconButton, StyledButtonRow, BottomAlignedTypography };
+export { Flex, LeftMarginButton, LeftMarginIconButton };
