+++ context/app/static/js/shared-styles/sections/RightAlignedButtonRow/RightAlignedButtonRow.jsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import styled from 'styled-components';
+import Typography from '@material-ui/core/Typography';
 
 import { Flex } from './style';
 
@@ -11,4 +13,14 @@ function RightAlignedButtonRow({ leftText, buttons, ...props }) {
   );
 }
 
+const StyledButtonRow = styled(RightAlignedButtonRow)`
+  margin-bottom: ${(props) => props.theme.spacing(0.5)}px;
+  min-height: 40px;
+`;
+
+const BottomAlignedTypography = styled(Typography)`
+  align-self: flex-end;
+`;
+
+export { StyledButtonRow, BottomAlignedTypography };
 export default RightAlignedButtonRow;
