+++ context/app/static/js/shared-styles/sections/RightAlignedButtonRow/index.js
@@ -1,3 +1,4 @@
-import RightAlignedButtonRow from './RightAlignedButtonRow';
+import RightAlignedButtonRow, { StyledButtonRow, BottomAlignedTypography } from './RightAlignedButtonRow';
 
+export { StyledButtonRow, BottomAlignedTypography };
 export default RightAlignedButtonRow;
